requirements.txt
# Poetry BERT Formalism - Python Dependencies
# Install with: pip install -r requirements.txt

# Core ML/NLP
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
accelerate>=0.20.0

# Word embeddings
gensim>=4.3.0

# Prosodic analysis
prosodic>=1.5.0
pronouncing>=0.2.0  # Phonetic rhyme detection

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Database
sqlite3  # Built-in with Python

# Progress tracking
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.23.0
ipywidgets>=8.0.0

# Google Drive integration (optional)
PyDrive2>=1.15.0

# Configuration management
python-dotenv>=1.0.0

# Testing (optional)
pytest>=7.3.0
pytest-cov>=4.1.0

# Code quality (optional, for development)
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
